pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autocue"
version = "0.1.0"
description = "Low-latency auto-cue teleprompter with local speech recognition"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "vosk>=0.3.42",
    "sounddevice>=0.4.6",
    "websockets>=12.0",
    "rapidfuzz>=3.6.0",
    "markdown>=3.5.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "num2words>=0.5.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
autocue = "autocue.main:main"
autocue-debug-transcript = "autocue.debug_transcript:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pylint.messages_control]
disable = [
    "W0613",  # unused-argument
    "W0718",  # broad-exception-caught
    "C0302",  # too-many-lines
]
